Added new project creation server calls
Clicking the create new project button on the project selector now brings up a form to fill in with the new project details.  Clicking the submit button sends the data to the server (currently a JSON file) and updates the Redux store with the details of the new project as the chosen project.

The screen does not update once the submit button is clicked.