Fixed logout, added project requests tab, fixed profile page
The logout was retaining information between users when someone logged out via the store; I added a function to reset the store to default when signing out.

The project request tab now generates tiles for each of the requests to join a project.  It does not have a "no requests" placeholder yet.

The user menu does not show how many requests a user has.

The requests page does not do any other types of request yet (sign document).

The profile page was structured in a poor way (business logic) causing memory leaks.  I moved all the calls to the profile page constructor to remedy this issue.  Lower down elements consume the data from the requests made by the profile page.